moves overlay to element.
